requirements.txt
anyio==4.9.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
Flask==3.1.0
Flask-Material==0.1.1
Flask-WTF==1.2.2
h11==0.14.0
hl7==0.4.5
httpcore==1.0.8
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pydicom==2.4.4
pymongo==4.11.3
pyorthanc==1.20.0
requests==2.32.3
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.3.0
Werkzeug==3.1.3
WTForms==3.2.1

numpy~=2.2.4
txHL7~=0.5.0
radgraph~=0.1.18